abm comprobantes

+++ back/src/repositories/comprobante_repo.py
@@ -1,4 +1,5 @@
 from sqlalchemy.orm import Session
+from fastapi import HTTPException
 
 from src.models.comprobante import Comprobante
 from src.schemas.comprobante_schema import ComprobanteCreate
@@ -21,8 +22,20 @@ class ComprobanteRepo:
         self.db.refresh(db_comprobante)
         return db_comprobante
     
+    def update_comprobante(self, id: int, comprobante: ComprobanteCreate):
+        db_comprobante = self.db.query(Comprobante).filter(Comprobante.id == id).first()
+        if db_comprobante is None:
+            raise HTTPException(status_code=404, detail="Comprobante no encontrado")
+        for key, value in comprobante.items():
+            setattr(db_comprobante, key, value)
+        self.db.commit()
+        self.db.refresh(db_comprobante)
+        return db_comprobante
+    
     def delete_comprobante(self, id: int):
         db_comprobante = self.db.query(Comprobante).filter(Comprobante.id == id).first()
+        if db_comprobante is None:
+            raise HTTPException(status_code=404, detail="Comprobante no encontrado")
         self.db.delete(db_comprobante)
         self.db.commit()
         return db_comprobante
